Renamed loginRouter to rootRouter
